+++ __tests__/sliceImageUseCaseOnFigma.spec.ts
@@ -52,7 +52,6 @@ describe('tests for sliceImageUseCase on Figma', () => {
   });
 
   afterEach(() => {
-    // await new Promise(resolve => setTimeout(resolve, 1000));
     process.env = OLD_ENV;
   });