+++ core/domain/Entities.ts
@@ -57,12 +57,10 @@ export { AssetFormat } from './entities/AssetFormat';
 export { SliceConfig } from './entities/SliceConfig';
 export {
   DtcConfig,
-  SketchConfig,
-  SymbolElement,
-  SketchExtraction,
 } from './entities/DtcConfig';
 
 export { GenerateConfig } from './entities/GenerateConfig';
 export { StyleConfig } from './entities/StyleConfig';
 export { ColorStyleConfig } from './entities/ColorStyleConfig';
 export { Styles } from './entities/Styles';
+export { StyleType } from './entities/StyleType';