+++ core/types.ts
@@ -1,17 +1,13 @@
 const TYPES = {
   INamingLinter: Symbol.for('INamingLinter'),
-  ISketchPresenter: Symbol.for('ISketchPresenter'),
-  ISketchRepository: Symbol.for('ISketchRepository'),
   IFigmaPresenter: Symbol.for('IFigmaPresenter'),
   IFigmaRepository: Symbol.for('IFigmaRepository'),
   IFigmaConfig: Symbol.for('IFigmaConfig'),
   IGenericUseCase: Symbol.for('IGenericUseCase'),
   IGenericRepository: Symbol.for('IGenericRepository'),
   ILintNamingUseCase: Symbol.for('ILintNamingUseCase'),
-  IExtractElementUseCase: Symbol.for('IExtractElementUseCase'),
   ISliceImageUseCase: Symbol.for('ISliceImageUseCase'),
   IStyleUseCase: Symbol.for('IStyleUseCase'),
-  IGenerateCodeUseCase: Symbol.for('IGenerateCodeUseCase'),
   IGenerateAssetUseCase: Symbol.for('IGenerateAssetUseCase'),
   IGenerateProjectUseCase: Symbol.for('IGenerateProjectUseCase'),
 };